Converted project into maven repo
