--- src/AttendanceRecord.java
@@ -1,13 +1,0 @@
-import java.sql.Time;
-import java.util.Date;
-
-public class AttendanceRecord {
-    private String name;
-    private String id;
-    private Date date;
-    private Time timeIn;
-    private Time timeOut;
-
-    public double calcHoursWorked;
-
-}
