--- src/EmployeeDetails.java
@@ -1,27 +1,0 @@
-public class EmployeeDetails {
-
-    private String name;
-    private String id;
-    private String address;
-    private String birthday;
-    private String status;
-    private String position;
-    private String phoneNumber;
-    private String sss;
-    private String philHealth;
-    private String tin;
-    private String pagibig;
-    private double hourlyRate;
-    private double grossSalary;
-    private double basicSalary;
-    private double riceSubsidy;
-    private double phoneAllowance;
-    private double clothingAllowance;
-
-public double calcSSS;
-public double calcPhilHealth;
-public double calcPagIbig;
-public double calcTax;
-
-
-}
