--- src/LogIn.java
@@ -1,5 +1,0 @@
-public class LogIn {
-    private double employeeId;
-    private String password;
-
-}
