+++ src/main/java/LogIn.java
@@ -1,0 +1,5 @@
+public class LogIn {
+    private double employeeId;
+    private String password;
+
+}
